refactor window creation to share code
