+++ lua/azuredo/init.lua
@@ -1,5 +1,6 @@
 local Config = require("azuredo.config")
 local Util = require("azuredo.util")
+local Window = require("azuredo.window")
 
 local M = {}
 
@@ -9,60 +10,24 @@ function M.setup(opts)
   require("azuredo.config").setup(opts)
 end
 
-function M.createWindow()
+function M.openMainMenu()
   local options = {
     "Create Pull Request",
     "Add Work item to Pull Request",
     "Print PR Id",
-    "Set PR Id",
+    "Set Existing PR Id",
+    "Set PR Id manually",
   }
 
-  local buf = vim.api.nvim_create_buf(false, true)
-  local width = 80
-  local max_height = 20
-  local height = math.min(#options + 2, max_height)
-
-  local ui = vim.api.nvim_list_uis()[1]
-
-  local opts = {
-    relative = "editor",
-    width = width,
-    height = height,
-    row = math.floor((ui.height - height) / 2),
-    col = math.floor((ui.width - width) / 2),
-    anchor = "NW",
-    style = "minimal",
-    border = "rounded",
-  }
-
-  local win = vim.api.nvim_open_win(buf, false, opts)
-  vim.api.nvim_set_current_win(win)
-  vim.api.nvim_buf_set_lines(buf, 0, -1, true, options)
-  vim.bo[buf].modifiable = false
-  vim.bo[buf].buftype = "nofile"
-  vim.api.nvim_set_option_value("winhl", "Normal:MyHighlight", { win = win })
-  vim.api.nvim_win_set_cursor(win, { 1, 0 })
-
-  local function select_current_line()
-    local row, _ = unpack(vim.api.nvim_win_get_cursor(win))
+  ---@param row integer
+  local function select_current_line(row)
     local choice = options[row]
     if choice then
-      vim.api.nvim_win_close(win, true)
       M.executeCommand(choice)
     end
   end
 
-  vim.keymap.set("n", "q", function()
-    vim.api.nvim_win_close(win, true)
-  end, { buffer = buf, silent = true })
-  vim.keymap.set("n", "<CR>", select_current_line, { buffer = buf, silent = true })
-
-  for i, option in ipairs(options) do
-    vim.keymap.set("n", tostring(i), function()
-      vim.api.nvim_win_close(win, true)
-      Util.debug(option)
-    end, { buffer = buf, silent = true })
-  end
+  Window.createWindow(options, select_current_line)
 end
 
 M.prId = nil
@@ -88,8 +53,19 @@ function M.executeCommand(command)
     M.fetch_and_show_workitems()
   elseif command == "Print PR Id" then
     print(M.prId)
-  elseif command == "Set PR Id" then
+  elseif command == "Set PR Id manually" then
     M.prId = vim.fn.input("Enter PR ID: ")
+  elseif command == "Set Existing PR Id" then
+    local result = vim.fn.system("az repos pr list --source-branch $(git branch --show-current) --output json")
+
+    local success, pr_data = pcall(vim.json.decode, result)
+    if success and pr_data and pr_data[1].pullRequestId then
+      print("Pull Request ID: " .. pr_data[1].pullRequestId)
+      M.prId = pr_data[1].pullRequestId
+    else
+      print("Failed to get ID. PR doesn't exist or something went wrong")
+      Util.debug(result)
+    end
   end
 end
 
@@ -133,54 +109,22 @@ function M.fetch_and_show_workitems()
     table.insert(work_item_ids, id)
   end
 
-  local buf = vim.api.nvim_create_buf(false, true)
-  local width = 80
-  local max_height = 20
-  local height = math.min(#work_items + 2, max_height)
-
-  local ui = vim.api.nvim_list_uis()[1]
-  local win_opts = {
-    relative = "editor",
-    width = width,
-    height = height,
-    row = math.floor((ui.height - height) / 2),
-    col = math.floor((ui.width - width) / 2),
-    style = "minimal",
-    border = "rounded",
-  }
-
-  local win = vim.api.nvim_open_win(buf, true, win_opts)
-  vim.api.nvim_set_current_win(win)
-
-  vim.api.nvim_buf_set_lines(buf, 0, -1, false, work_items)
-
-  vim.bo[buf].modifiable = false
-  vim.bo[buf].buftype = "nofile"
-  vim.bo[buf].bufhidden = "wipe"
-  vim.bo[buf].swapfile = false
-
-  vim.api.nvim_win_set_cursor(win, { 1, 0 })
-
-  local function select_current_line()
-    local row, _ = unpack(vim.api.nvim_win_get_cursor(win))
+  ---@param row integer
+  local function select_current_line(row)
     local selected_id = work_item_ids[row]
     if selected_id then
-      vim.api.nvim_win_close(win, true)
       print("Adding Id" .. selected_id .. " to PR " .. M.prId)
       local work_item_cmd = [[az repos pr work-item add --id ]] .. M.prId .. [[ --work-items ]] .. selected_id
       vim.fn.system(work_item_cmd)
     end
   end
 
-  vim.keymap.set("n", "q", function()
-    vim.api.nvim_win_close(win, true)
-  end, { buffer = buf, silent = true })
-  vim.keymap.set("n", "<CR>", select_current_line, { buffer = buf, silent = true })
+  Window.createWindow(work_items, select_current_line)
 end
 
 vim.api.nvim_create_user_command("Azuredo", function(opts)
   if opts.args then
-    M.createWindow()
+    M.openMainMenu()
     return
   end
 end, { nargs = "*", desc = "azuredo plugin" })
